Revert "Refacor router (#5)"
This reverts commit 24ededbcc21ee1cff32e0388e663e49ec20d31a2.
